% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywoe.R
\name{add_woe}
\alias{add_woe}
\title{Add WoE in a data.frame}
\usage{
add_woe(.data, .response, ..., .woe_dictionary = NULL)
}
\arguments{
\item{.data}{A tbl. The data.frame to plug the new woe version columns.}

\item{.response}{unquoted name of the response variable.}

\item{...}{unquoted names of explanatory variables, passed as you would pass variables to \code{dplyr::select()}. This means that you can use all the helpers like \code{starts_with()} and \code{matches()}.}
}
\value{
a tibble with the original columns of .data plus the woe columns wanted.
}
\description{
A tidyverse friendly way to plug WoE versions of a set of explanatory variables against a given dicotomous response.
}
\details{
You can pass a custom dictionary to \code{add_woe()}. It must have the exactly the same structure of the output of \code{woe_dictionary()}. One easy way to do this is to tweak a output returned from it.
}
\examples{

mtcars \%>\% add_woe(am, cyl, gear:carb)

}
