% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywoe.R
\name{woe_dictionary}
\alias{woe_dictionary}
\title{WoE dictionary of a set of explanatory variables upon a given dicotomous response}
\usage{
woe_dictionary(.data, .response, ...)
}
\arguments{
\item{.data}{A tbl. The data.frame where the variables come from.}

\item{.response}{unquoted name of the response variable.}

\item{...}{unquoted names of explanatory variables, passed as you would pass variables to \code{dplyr::select()}. This means that you can use all the helpers like \code{starts_with()} and \code{matches()}.}
}
\value{
a tibble with summaries and woe for every given explanatory variable stacked up. Warning: WoE can possibly be -Inf.
}
\description{
Builds the WoE dictionary of a set of explanatory variables upon a given dicotomous response.
Convenient to make a WoE version of the given set of explanatory variables and also to allow
one to tweak some woe values by hand.
}
\details{
You can pass a custom dictionary to \code{add_woe()}. It must have the exactly the same structure of the output of \code{woe_dictionary()}. One easy way to do this is to tweak a output returned from it.
}
\examples{

mtcars \%>\% woe_dictionary(am, cyl, gear:carb)

}
