% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywoe.R
\name{woe_table}
\alias{woe_table}
\title{Cross table with woe between a (dicotomous) response and a explanatory}
\usage{
woe_table(response, explanatory)
}
\arguments{
\item{response}{The dependent variable. A atomic vector with exactly 2 distinct values.}

\item{explanatory}{A atomic vector, usualy with few distinct values.}
}
\value{
a tibble with summaries and woe. Warning: WoE can possibly be -Inf.
}
\description{
Calculates some summaries and the WoE between a dicotomous response and a given explanatory variable.
Used to biuld the dictionary.
}
\examples{

resp <- rep(c("A", "B"), 20)
expl <- sample(c("X", "Y", "Z", "W"), size = 40, replace = TRUE)
woe_table(resp, expl)

}
